+++ steelc/src/building/build_config.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+
+class build_config {
+public:
+	// this constructor creates a default build configuration
+	build_config() = default;
+
+	// output options
+	std::string output_dir = "build/";
+	std::string intermediate_dir = "build/cache/";
+	std::string build_cache_file = "build/cache/build_cache.dat";
+};