+++ steelc/src/input/commands/steelc_command.cpp
@@ -23,12 +23,26 @@ bool steelc_command::execute(const console_args& args) const {
 		return false;
 	}
 
+	// handle global flags
+	if (flags.has("--verbose")) {
+		output::set_log_verbosity(LOG_VERBOSITY_HIGH);
+	}
+
 	return handler(flags);
+
+	if (flags.has("--verbose")) {
+		output::set_log_verbosity(LOG_VERBOSITY_NORMAL);
+	}
 }
 
 bool steelc_command::parse_args(const console_args& args, command_flags& out_flags) const {
 	std::unordered_set<std::string> seen_flags;
 	std::unordered_map<std::string, std::vector<std::string>> parsed_flags;
+
+	// should probably move this into a dedicated function or something
+	static const std::unordered_set<std::string> global_flags = {
+		"--verbose"
+	};
 	
 	const auto& cfg_flags = flag_cfg->get_flags();
 	const auto& req_values = flag_cfg->get_required_values();
@@ -46,7 +60,9 @@ bool steelc_command::parse_args(const console_args& args, command_flags& out_fla
 	while (arg_it != args.end()) {
 		const std::string& arg = *arg_it;
 
-		if (!cfg_flags.contains(arg)) {
+		bool is_global = global_flags.contains(arg);
+
+		if (!cfg_flags.contains(arg) && !is_global) {
 			output::print("Warn: unknown flag: {}\n", console_colors::YELLOW, arg);
 			++arg_it;
 			continue;
@@ -55,6 +71,13 @@ bool steelc_command::parse_args(const console_args& args, command_flags& out_fla
 		if (seen_flags.contains(arg)) {
 			output::err("Error: duplicate flag: {}\n", console_colors::RED, arg);
 			return false;
+		}
+
+		if (is_global) {
+			parsed_flags[arg] = {};
+			seen_flags.insert(arg);
+			++arg_it;
+			continue;
 		}
 
 		int expected = cfg_flags.at(arg);