+++ steelc/src/stproj/source_file.h
@@ -2,12 +2,17 @@
 
 #include <string>
 #include <vector>
+#include <cstdint>
 
 class source_file {
 public:
 	source_file(const std::string& full_path, const std::string& relative_path);
 
 	std::string name() const;
+
+	// filesystem utilites
+	uint64_t get_last_modified_time() const;
+	uint64_t get_size() const;
 
 	std::string relative_path;
 	std::string full_path;